requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision
torchaudio

# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0

# Tokenization
sentencepiece>=0.1.99

# Progress Tracking
tqdm>=4.65.0

# Metrics
rouge-score>=0.1.2

# Config Management
PyYAML>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional utilities
scikit-learn>=1.3.0
